backend/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.4
django-cors-headers==3.11.0
django-debug-toolbar==3.4.0
django-filter==21.1
django-oauth-toolkit==2.0.0
djangorestframework==3.13.1
drf-social-oauth2==1.2.1
ecdsa==0.17.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.1
jwcrypto==1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
Pillow==9.1.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.20.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.1
